adds basic camera controls with WASD+QE+CTRL

+++ source/frame_info.hpp
@@ -6,5 +6,5 @@ struct FrameInfo
 {
 	u64 idx;
 	f64 seconds_since_start;
-	f64 seconds_since_last_frame;
+	f32 seconds_since_last_frame;
 };
